requirements.txt
plaid-python>=3.0.0
Flask>=1.1.1
pyAesCrypt>=0.4.3
money>=1.3
babel>=2.8
pendulum>=2.0.5
